+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## HEAD
+
+* Add `TWEnv::Command::PerformActionOnTweets`.
+
 ## v0.2.0
 
 * Fix error when falling through to `Twitter::Errors::TooManyRequests`.